pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["Deminform <**********@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
aiohttp = "^3.11.7"
python-dotenv = "^1.0.1"
certifi = "^2024.8.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
